fixed typos

+++ algorithms/mex_interfaces/taukl_cc_mex_interface.m
@@ -1,6 +1,6 @@
 function [taukl_val] = taukl_cc_mex_interface(X,Y,autoDetectHybrid,isHybrid,continuousRvIndicator)
 
 [u,v] = pobs_sorted_cc_mex(X,Y); 
-taukl_val = taukl_cc_mex(u,v,int32(autoDetectHybrid),int32(isHybrid),int32(continuousRvIndicator));
+taukl_val = taukl_cc_mex(u,v,autoDetectHybrid,isHybrid,continuousRvIndicator);
 
 end
